// client/app/api/accounting/model/EmployeeModel.ts
/**
 * GuanPlus.AccountingFirm.WebApi
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import * as models from './models';

/**
 * 员工model
 */
export interface EmployeeModel {
    /**
     * id
     */
    id?: string;

    /**
     * 编号
     */
    number?: string;

    /**
     * 名
     */
    name?: string;

    /**
     * 职位
     */
    position?: string;

    /**
     * 身份证号
     */
    idNo?: string;

    /**
     * 年龄
     */
    birthDate?: Date;

    /**
     * 入职时间
     */
    joinedDate?: Date;

    /**
     * 国籍 id为code name为名
     */
    national?: models.IdNameModel;

    /**
     * 部门类型 0 无 1 销售 2 管理
     */
    departmentType?: models.DepartmentTypeEnumModel;

    /**
     * 社会保险
     */
    hasSocialSecurity?: boolean;

    /**
     * 性别 无0 男1 女2
     */
    gender?: models.GenderEnumModel;

    /**
     * 股权 个人股本（投资）额？
     */
    shareCapital?: number;

    /**
     * 工作单位
     */
    companyName?: string;

    /**
     * 户口性质 - 农业1 非农业2
     */
    accountNature?: models.PersonalAccountTypeEnumModel;

    /**
     * 
     */
    phoneNumber?: string;

    /**
     * 
     */
    email?: string;

    /**
     * 是否为孤老
     */
    remainsElderly?: string;

    /**
     * 住房公积金
     */
    hasHousingFund?: boolean;

    /**
     * 状态 正常0 离职1
     */
    status?: models.EmployeeStatusEnumModel;

    /**
     * 合同类型-是否为雇员
     */
    contractType?: string;

    /**
     * 地址
     */
    address?: string;

    /**
     * 基本工资
     */
    baseSalary?: number;

    /**
     * 岗位工资
     */
    positionSalary?: number;

    /**
     * 开户银行
     */
    bankType?: string;

    /**
     * 银行账号
     */
    bankAccount?: string;

    /**
     * 医疗保险基数
     */
    medicalInsuranceBase?: number;

    /**
     * 养老保险基数
     */
    oldAgeInsuranceBase?: number;

    /**
     * 公积金基数
     */
    providentFundBase?: number;

    /**
     * 个人医疗保险
     */
    pEdicalCare?: number;

    /**
     * 个人养老保险
     */
    pPension?: number;

    /**
     * 个人失业保险
     */
    pUnemployment?: number;

    /**
     * 个人公积金
     */
    pProvidentFund?: number;

    /**
     * 公司医疗保险
     */
    cEdicalCare?: number;

    /**
     * 公司养老保险
     */
    cPension?: number;

    /**
     * 公司失业保险
     */
    cUnemployment?: number;

    /**
     * 公司公积金
     */
    cProvidentFund?: number;

    /**
     * 公司生育
     */
    cFertility?: number;

    /**
     * 公司工伤
     */
    cWorkInjury?: number;

    /**
     * 导入excel中对应的行数
     */
    importRow?: number;

    /**
     * ImportTaskId
     */
    importTaskId?: string;

}